od5.0: Fixed conversion errors
Summary: Fixed conversion errors

Test Plan: According to CDash errors are gone

Reviewers: yuancheng

Reviewed By: yuancheng

Differential Revision: https://phabricator.dgbes.com:8443/D3970

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@34730 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Algo/axislayout.h
@@ -75,7 +75,7 @@ void AxisLayout<T>::setDataRange( const Interval<T>& dr )
 	return;
     }
 
-    double scwdth = wdth < 1e-30 ? -30 : log10( wdth );
+    double scwdth = wdth < 1e-30 ? -30 : log10( (double)wdth );
     int tenpow = 1 - (int)scwdth; if ( scwdth < 0 ) tenpow++;
     double stepfac = Math::IntPowerOf( ((double)10), tenpow );
     scwdth = wdth * stepfac;
